% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branched_pt.R
\name{branched_pt}
\alias{branched_pt}
\title{Branched pt function}
\usage{
branched_pt(env, root, tips)
}
\arguments{
\item{env}{An enviroment produced by the oposSOM pipeline}

\item{root}{A single group label defined as the developmental source}

\item{tips}{A vector of group labels defined as the developmental sinks}
}
\description{
This function reconstructs transcriptional trajectories between a developmental root and the tips.
}
\keyword{urd}
