% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_run.R
\name{pt_run}
\alias{pt_run}
\title{oposSOM PT run Function}
\usage{
pt_run(env, root = c(), tips = c(), lineages = list(), loom_path = c())
}
\arguments{
\item{env}{An enviroment produced by the oposSOM pipeline.}

\item{root}{A single group label defined as the developmental source.}

\item{tips}{A vector of group labels defined as the developmental sinks.}

\item{lineages}{A list of vectors of group labels sorted by pseudotime.}
}
\description{
This function allows you to run the oposSOM pseudotemporal analysis.
}
