% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_gradients.R
\name{som_gradients}
\alias{som_gradients}
\title{SOM gradients function}
\usage{
som_gradients(env, root = c(), tips = c(), lineages = c())
}
\arguments{
\item{env}{An enviroment produced by the oposSOM pipeline.}

\item{root}{A single group label defined as the developmental source.}

\item{tips}{A vector of group labels defined as the developmental sinks.}

\item{lineages}{A list of vectors of group labels sorted by pseudotime.}
}
\description{
This function provides the vector field (gradient field) and the corresponding stream lines in expression portraits.
}
