% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_lcp.R
\name{som_lcp}
\alias{som_lcp}
\title{SOM Least Cost Path (LCP) function}
\usage{
som_lcp(env, root = c(), tips = c(), lineages = c())
}
\arguments{
\item{env}{An enviroment produced by the oposSOM pipeline.}

\item{root}{A single group label defined as the developmental source.}

\item{tips}{A vector of group labels defined as the developmental sinks.}

\item{lineages}{A list of vectors of group labels sorted by pseudotime.}
}
\description{
This function calculate the difference portraits between the root and the tips
to generate gradient trajectories along topocrafic least cost paths
}
\keyword{LCP}
\keyword{distance}
\keyword{topografic}
