% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_mep.R
\name{som_mep}
\alias{som_mep}
\title{SOM Maximum Expression Path (MEP) Function}
\usage{
som_mep(env, root = c(), tips = c(), lineages = c())
}
\arguments{
\item{env}{An enviroment produced by the oposSOM pipeline.}

\item{root}{A single group label defined as the developmental source.}

\item{tips}{A vector of group labels defined as the developmental sinks.}

\item{lineages}{A list of vectors of group labels sorted by pseudotime.}
}
\description{
This function connects root and tips along the SOM clusters providing the maximum cumulative weights
and then defines the gene state trajectory ensuring maximum cumulative expression between source and
sink
}
